moved cluster access views to separate module
